Added Map Functionality.
Players must clear one zone before moving on to the others. Zones that
have been fully cleared become green and can be revisited. Zones that
are locked are grey. Zones that you should currently be questing in to
advance appear white.

Added Animation so you zoom into the zone you’re going to be questing
in.

I will be adding a player to reinstate my code that saved skins, and
set the skin you selected on to the character, wherever they appeared.
Hope to add this tmrw (Thursday)